requirements.txt
Django==2.0.2
django-webpack-loader==0.5.0
djangorestframework==3.7.7
flake8==3.5.0
mccabe==0.6.1
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.3
